+++ ariadne/schema_names.py
@@ -1,4 +1,4 @@
-from typing import Callable, Optional, Type, Union
+from typing import Callable, Optional, Tuple, Union
 
 from graphql import (
     GraphQLArgument,
@@ -13,7 +13,15 @@ from .resolvers import resolve_to
 from .utils import convert_camel_case_to_snake
 
 SchemaNameConverter = Callable[
-    [str, Type[Union[GraphQLArgument, GraphQLField, GraphQLInputField]]], str
+    [
+        str,
+        Union[
+            Tuple[GraphQLObjectType, GraphQLField],
+            Tuple[GraphQLObjectType, GraphQLField, GraphQLArgument],
+            Tuple[GraphQLInputObjectType, GraphQLInputField],
+        ],
+    ],
+    str,
 ]
 
 GRAPHQL_SPEC_TYPES = (
@@ -43,40 +51,41 @@ def convert_schema_names(
 
 
 def convert_names_in_schema_object(
-    graphql_type: GraphQLObjectType,
+    graphql_object: GraphQLObjectType,
     name_converter: SchemaNameConverter,
 ) -> None:
-    for field_name, field in graphql_type.fields.items():
+    for field_name, field in graphql_object.fields.items():
         if field.args:
-            convert_names_in_schema_args(field, name_converter)
+            convert_names_in_schema_args(graphql_object, field, name_converter)
 
         if field.resolve or field_name.lower() == field_name:
             continue
 
-        field.resolve = resolve_to(name_converter(field_name, GraphQLField))
+        field.resolve = resolve_to(name_converter(field_name, (graphql_object, field)))
 
 
 def convert_names_in_schema_args(
-    graphql_type: GraphQLField,
+    graphql_object: GraphQLObjectType,
+    graphql_field: GraphQLField,
     name_converter: SchemaNameConverter,
 ) -> None:
-    for arg_name, arg in graphql_type.args.items():
+    for arg_name, arg in graphql_field.args.items():
         if arg.out_name or arg_name.lower() == arg_name:
             continue
 
-        arg.out_name = name_converter(arg_name, GraphQLInputField)
+        arg.out_name = name_converter(arg_name, (graphql_object, graphql_field, arg))
 
 
 def convert_names_in_schema_input(
-    graphql_type: GraphQLInputObjectType,
+    graphql_input: GraphQLInputObjectType,
     name_converter: SchemaNameConverter,
 ) -> None:
-    for field_name, field in graphql_type.fields.items():
+    for field_name, field in graphql_input.fields.items():
         if field.out_name or field_name.lower() == field_name:
             continue
 
-        field.out_name = name_converter(field_name, GraphQLInputField)
+        field.out_name = name_converter(field_name, (graphql_input, field))
 
 
-def default_schema_name_converter(graphql_name: str, _graphql_type) -> str:
+def default_schema_name_converter(graphql_name: str, _path) -> str:
     return convert_camel_case_to_snake(graphql_name)